requirements.txt
dbt-snowflake==1.0.0
sqlfluff-templater-dbt==0.11.1
python-dotenv==0.19.2
snowflake-connector-python==2.8.0
